Fix splitDateTime function to work with IE7
The syntax for the second part of the splice does not work in IE7 (possibly others I haven't checked).  Have changed to a zero start index without reversing the sign of the timePartsLen.  As splice removes elements from the array, this way makes more logical sense anyway.  Works in IE7, IE9 and chrome 21.0. 