Fix issue with lowercase i18n am selector format
Lowercase am selector formats are interpreted as PM because of an error in the am selector recognition logic, line 1238, jquery-ui-timepicker-addon.js:
Original:
						ampm = $.inArray(treg[order.t].toUpperCase(), o.amNames) !== -1 ? 'AM' : 'PM';
Proposed:
						ampm = $.inArray(treg[order.t].toUpperCase(), o.amNames.toUpperCase()) !== -1 ? 'AM' : 'PM';

This bug causes lowercase am selector strings in amNames to be interpreted as PM when parsing the original value, resulting in inadvertent change of, for instance, 10 a.m. to 10 p.m. If the user didn't intend to change the date/time in the first place this is specially damaging because the user can miss the change and commit the form with the wrong time. Also note that amNames[0], which is used when wring back to the field is usually in lowercase.
Fix: Using the same transformation (toUpperCase) in both sides preserves the equality.